Fixed vod/ceta, added edit and delete options, fixed creation of new odred (added first vod and first ceta), users can't be deleted if holding a function. Added profile picture import system for odred, changed default profile picture for odred. Fixed horizontal scroll

+++ routes/odred.py
@@ -27,8 +27,9 @@ def odredDashboard(id):
         return redirect(url_for("dashboard.dashboard"))
     cetas = Ceta.query.filter_by(odred_id=current_user.odred_id).all()
     ceta_ids = [ceta.id for ceta in cetas]
-    vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).all()
-    return render_template("odredDashboard.html", odred_name=Odred.query.filter_by(id=id).first(), users=User.query.filter_by(odred_id=id).all(), vods = vods)
+    vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).order_by(Vod.ceta_id).all()
+    return render_template("odredDashboard.html", odred_name=Odred.query.filter_by(id=id).first(),
+                           users=User.query.filter_by(odred_id=id).all(), vods=vods)
 
 
 # Dodavanje novog clana u odred
@@ -39,13 +40,12 @@ def addClan():
     new_user = User()
     if request.method == "POST":
         new_user = User.defUser(new_user)
-
         new_user.set_password(new_user.username)
-        new_user.odred_id = current_user.odred_id
-
         db.session.add(new_user)
         db.session.commit()
         new_user.username = f"{new_user.username}{new_user.id}c"
+        new_user.set_password(new_user.username)
+        new_user.odred_id = current_user.odred_id
         db.session.commit()
         return redirect(url_for("odred.odredDashboard", id=current_user.odred.id))
     else:
@@ -64,19 +64,21 @@ def addClan():
 def editClan(id):
     action = "edit"
     user = User.query.get(id)
+    cetas = Ceta.query.filter_by(odred_id=current_user.odred_id).all()
+    ceta_ids = [ceta.id for ceta in cetas]
+    vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).all()
     if request.method == "POST":
+        if user.vod.vodnik_id == user.id and request.form["vod"] != user.vod.name:
+            flash("Clan je vodnik svog voda! Prvo postavi drugog vodnika.", "error")
+            return render_template("addClan.html", h1="Izmeni člana", action=action, clan=user,
+                                   odred=Odred.query.filter_by(id=user.odred_id).first().name, vods=vods)
         user = User.defUser(user)
-
         db.session.commit()
         return redirect(url_for("odred.odredDashboard", id=current_user.odred.id))
     else:
         # Mogu da pristupe samo admini
         if current_user.role != "admin" and current_user.role != "savez_admin":
             return redirect(url_for("dashboard.dashboard"))
-
-        cetas = Ceta.query.filter_by(odred_id=current_user.odred_id).all()
-        ceta_ids = [ceta.id for ceta in cetas]
-        vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).all()
         return render_template("addClan.html", h1="Izmeni člana", action=action, clan=user,
                                odred=Odred.query.filter_by(id=user.odred_id).first().name, vods=vods)
 
@@ -91,9 +93,12 @@ def deleteClan(id):
 
     user = User.query.get(id)
     if user:
-        db.session.delete(user)
-        db.session.commit()
-        flash("Član je uspešno obrisan.", "Info")
+        if user.odred.staresina_id != user.id and user.odred.nacelnik_id != user.id and user.vod.vodnik_id != user.id and user.vod.ceta.vodja_id != user.id:
+            db.session.delete(user)
+            db.session.commit()
+            flash("Član je uspešno obrisan.", "Info")
+        else:
+            flash("Član vrši dužnost u odredu! Možeš obrisati samo članove koji su razrešeni dužnosti", "error")
     else:
         flash("Član nije pronađen.", "Greška")
     return redirect(url_for("odred.odredDashboard", id=current_user.odred.id))
@@ -119,94 +124,155 @@ def getPfp(id):
 @odred_bp.route("/cv/add", methods=["GET", "POST"])
 @login_required
 def addCetaVod():
-    return render_template("addCetaVod.html", users=User.query.filter_by(odred_id=current_user.odred_id).all(),
-                           cetas=Ceta.query.filter_by(odred_id=current_user.odred_id).all())
+    if current_user.role == "admin":
+        return render_template("addCetaVod.html", users=User.query.filter_by(odred_id=current_user.odred_id).all(),
+                               cetas=Ceta.query.filter_by(odred_id=current_user.odred_id).all())
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/ceta/new", methods=["POST"])
 @login_required
 def newCeta():
-    ceta = Ceta()
+    if current_user.role == "admin":
+        ceta = Ceta()
 
-    ceta.name = request.form["name"]
-    cete = Ceta.query.filter_by(odred_id=current_user.odred.id).all()
-    if any(existing_ceta.name == ceta.name for existing_ceta in cete):
-        flash("Četa sa tim imenom već postoji u odredu!", "Greška")
+        ceta.name = request.form["name"]
+        cete = Ceta.query.filter_by(odred_id=current_user.odred.id).all()
+        if any(existing_ceta.name == ceta.name for existing_ceta in cete):
+            flash("Četa sa tim imenom već postoji u odredu!", "Greška")
+            return redirect(url_for("odred.addCetaVod"))
+        ceta.odred_id = current_user.odred.id
+        db.session.add(ceta)
+        db.session.commit()
+
+        flash("Četa uspešno dodata!", "Info")
         return redirect(url_for("odred.addCetaVod"))
-    db.session.add(ceta)
-    db.session.commit()
-
-    flash("Četa uspešno dodata!", "Info")
-    return redirect(url_for("odred.addCetaVod"))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/vod/new", methods=["POST"])
 @login_required
 def newVod():
-    vod = Vod()
+    if current_user.role == "admin":
+        vod = Vod()
 
-    vod.name = request.form["name"]
-    vod.vodnik_id = request.form["vodnik"]
-    vod.ceta_id = request.form["ceta"]
+        vod.name = request.form["name"]
+        vod.vodnik_id = request.form["vodnik"]
+        vod.ceta_id = request.form["ceta"]
 
-    db.session.add(vod)
-    db.session.commit()
+        db.session.add(vod)
+        db.session.commit()
 
-    flash("Vod uspešno dodat!", "Info")
-    return redirect(url_for("odred.addCetaVod"))
+        vod.vodnik.vod_id = vod.id
+        db.session.commit()
+
+        flash("Vod uspešno dodat!", "Info")
+        return redirect(url_for("odred.addCetaVod"))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/edit/roles")
 @login_required
 def editRoles():
-    cetas = Ceta.query.filter_by(odred_id=current_user.odred_id).all()
-    ceta_ids = [ceta.id for ceta in cetas]
-    vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).all()
-    users = User.query.filter_by(odred_id=current_user.odred_id).order_by(User.dob.asc()).all()
+    if current_user.role == "admin":
+        cetas = Ceta.query.filter_by(odred_id=current_user.odred_id).all()
+        ceta_ids = [ceta.id for ceta in cetas]
+        vods = Vod.query.filter(Vod.ceta_id.in_(ceta_ids)).all()
+        users = User.query.filter_by(odred_id=current_user.odred_id).order_by(User.dob.asc()).all()
 
-    return render_template("editRoles.html", vods=vods, users=users, cetas=cetas,
-                           current_staresina=current_user.odred.staresina.id,
-                           current_nacelnik=current_user.odred.nacelnik.id)
+        return render_template("editRoles.html", vods=vods, users=users, cetas=cetas,
+                               current_staresina=current_user.odred.staresina.id,
+                               current_nacelnik=current_user.odred.nacelnik.id)
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/edit/roles/vodnik", methods=["POST"])
 @login_required
 def editVodnik():
-    vod = Vod.query.filter_by(id=request.form["vod"]).first()
+    if current_user.role == "admin":
+        vod = Vod.query.filter_by(id=request.form["vod"]).first()
 
-    vod.vodnik_id = request.form["vodnik"]
+        vod.vodnik_id = request.form["vodnik"]
 
-    db.session.commit()
+        db.session.commit()
 
-    flash("Novi vodnik postavljen", "Info")
-    return redirect(url_for("odred.editRoles"))
+        flash("Novi vodnik postavljen", "Info")
+        return redirect(url_for("odred.editRoles"))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/edit/roles/vodjacete", methods=["POST"])
 @login_required
 def editVodjacete():
-    ceta = Ceta.query.filter_by(id=request.form["ceta"]).first()
+    if current_user.role == "admin":
+        ceta = Ceta.query.filter_by(id=request.form["ceta"]).first()
 
-    ceta.vodja_id = request.form["vodjacete"]
+        ceta.vodja_id = request.form["vodjacete"]
 
-    db.session.commit()
+        db.session.commit()
 
-    flash("Novi vođa čete postavljen", "Info")
-    return redirect(url_for("odred.editRoles"))
+        flash("Novi vođa čete postavljen", "Info")
+        return redirect(url_for("odred.editRoles"))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
 
 
 @odred_bp.route("/edit/nacelnik", methods=["POST"])
 @login_required
 def editNacelnikStaresina():
-    current_user.odred.staresina_id = request.form["staresina"]
-    current_user.odred.nacelnik_id = request.form["nacelnik"]
+    if current_user.role == "admin":
+        current_user.odred.staresina_id = request.form["staresina"]
+        current_user.odred.nacelnik_id = request.form["nacelnik"]
 
-    db.session.commit()
+        db.session.commit()
 
-    flash("Načelnik i starešina uspešno ažurirani", "Info")
-    return render_template("editRoles.html")
+        flash("Načelnik i starešina uspešno ažurirani", "Info")
+        return redirect(url_for("odred.editRoles"))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
+
 
 @odred_bp.route("/vodInfo/<int:id>")
 @login_required
 def vodInfo(id):
-    return render_template("editVod.html",vod=Vod.query.filter_by(id=id).first(), users = User.query.filter_by(vod_id=id))
+    if current_user.role == "admin":
+        return render_template("editVod.html", vod=Vod.query.get(id),
+                               users=User.query.filter_by(vod_id=id),
+                               cete=Ceta.query.filter_by(odred_id=current_user.odred.id).all())
+    else:
+        return redirect(url_for("dashboard.dashboard"))
+
+
+@odred_bp.route("/editVod/<int:id>", methods = ["POST"])
+@login_required
+def editVod(id):
+    if current_user.role == "admin":
+        vod = Vod.query.get(id)
+        vod.name = request.form.get("vod_name")
+        vod.ceta_id = request.form.get("ceta")
+        db.session.commit()
+        return redirect(url_for("odred.vodInfo", id=id))
+    else:
+        return redirect("dashboard.dashboard")
+
+
+@odred_bp.route("/deleteVod/<int:id>")
+@login_required
+def deleteVod(id):
+    if current_user.role == "admin":
+        if len(User.query.filter_by(vod_id=id).all()) == 0:
+            vod = Vod.query.get(id)
+            db.session.delete(vod)
+            db.session.commit()
+            flash("Vod je uspešno obrisan", "info")
+            return redirect(url_for("odred.odredDashboard", id=current_user.odred_id))
+        else:
+            flash("Vod ne sme imati članove koji mu pripadaju ukoliko želite da ga obrišete", "error")
+            return redirect(url_for("odred.vodInfo", id=id))
+    else:
+        return redirect(url_for("dashboard.dashboard"))
